api/match-service/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.23
Werkzeug==3.1.3
psycopg2-binary==2.9.9
numpy==1.26.4
